Vendor import ntp-4.2.8p1.
Approved by:	roberto


Former-commit-id: b685303df7a56085fd00cafb6ca4acfc77a5d47d

+++ util/jitter.h
@@ -461,7 +461,7 @@ name|x_i
 parameter_list|,
 name|x_f
 parameter_list|)
-value|(((x_i)>= 0x00010000) ? 0x7fffffff : \ 				(((x_i)<= -0x00010000) ? 0x80000000 : \ 				(((x_i)<<16) | (((x_f)>>16)&0xffff))))
+value|(((int)(x_i)>= 0x00010000) ? 0x7fffffff : \ 				(((int)(x_i)<= -0x00010000) ? 0x80000000 : \ 				(u_int)(((x_i)<<16) | (((x_f)>>16)&0xffff))))
 end_define
 
 begin_define
